test(core,extension-block-menu): move the floating menu plugin tests into core and cover the placeholder and isEditable patch branches
